requirements.txt
pillow
pandas  
